fix(workrooms): complete overseer closeout (BI-3913EB49)
