+++ AGENTS.md
@@ -18,7 +18,7 @@ Every rule here is a one-line statement; the kernel principle behind it carries 
 - **Platform function never depends on a client.** Every guarantee runs server-side, on every install, with no AI client present; a hook may accelerate one, never own it. **Commandment tier.** → [kernel principle](docs/founder-kernel/wiki/principles/platform-function-never-depends-on-a-client.md)
 - **Single source of truth.** Each rule, fact or decision in exactly one place. Pointers, not copies. → [kernel principle](docs/founder-kernel/wiki/principles/single-source-of-truth.md)
 - **Ground new work in existing platform work.** Inspect the specs, schema, epics, principles, routes, primitives and backlog first; extend or refactor what exists. Net-new substrate only when prior work is proven unfit and the supersession explicit. → [kernel](docs/founder-kernel/wiki/principles/consult-specs-first.md) · [epics](docs/professions/portfolio-management/wiki/check-epic-overlap-before-creating.md) · [schema](docs/professions/data-architect/wiki/schema-audit-before-features.md) · [substrate](docs/founder-kernel/wiki/principles/verify-substrate-before-proposing-new.md)
-- **An enforcement refusal is a stop, not a workaround.** Branch protection, the DCO bot, token grant-intersection and — where the host loads them — `PreToolUse` hooks refuse rather than warn. Surface the requirement and stop; never route around it via DB, filesystem or shell. ⟦situational: hooks are Claude-format — native on Claude Code/Grok, aliased on Codex, unproven elsewhere; on an unverified host honour the hook-enforced rules yourself — review at EP-ANTIGRAVITY-001's evidence gate⟧ → [PR](docs/founder-kernel/wiki/principles/all-changes-land-via-pr.md) · [DCO](docs/founder-kernel/wiki/principles/dco-sign-off-required.md) · [lease](docs/founder-kernel/wiki/principles/runtime-gates-via-shared-lease.md)
+- **An enforcement refusal is a stop unless the gate already exposes a checked-in, audit-recording override and the operator explicitly authorizes that override for this exact run.** Use only that supported path, record its reason and scope, and report the gate as **skipped/unrun — never passed**. No override may bypass branch protection, DCO, token grant-intersection, destructive-safety controls, or production-integrity controls; never route around a refusal through direct DB, filesystem or shell mutation. If no supported override exists, surface the requirement and stop. ⟦situational: hooks are Claude-format — native on Claude Code/Grok, aliased on Codex, unproven elsewhere; on an unverified host honour the hook-enforced rules yourself — review at EP-ANTIGRAVITY-001's evidence gate⟧ → [PR](docs/founder-kernel/wiki/principles/all-changes-land-via-pr.md) · [DCO](docs/founder-kernel/wiki/principles/dco-sign-off-required.md) · [lease](docs/founder-kernel/wiki/principles/runtime-gates-via-shared-lease.md)
 - **The canonical runtime is the only source of runtime truth.** A hand-built image proves nothing about the live system. Runtime-bound verification, release validation and install advances route through the canonical runtime or shared local-CI lease; the live install advances only via `/ops/self-upgrade` (hand-tagging the image is hook-refused); an image carries the identity of its bytes. → [worktree](docs/founder-kernel/wiki/principles/worktree-is-source-control-not-runtime.md) · [release QA](docs/professions/release-service-management/wiki/release-qa-plan.md) · [image identity](docs/founder-kernel/wiki/principles/image-identity-equals-bytes.md)
 - **Architecture over shortcuts.** A quick fix that bypasses the design creates more debt than it saves. → [kernel principle](docs/founder-kernel/wiki/principles/architecture-over-shortcuts.md)
 - **Classify ambiguous requests before acting.** When a request could mean more than one work type, stop before code edits and have the operator classify it; prefer the highest-governance reading. → [kernel principle](docs/founder-kernel/wiki/principles/classify-ambiguous-requests-before-acting.md)
