+++ apps/web/components/ops/workrooms/WorkroomInventory.test.tsx
@@ -8,8 +8,8 @@ describe("WorkroomInventory", () => {
     const html = renderToStaticMarkup(<WorkroomInventory
       summary={{ scanned: 3, live: 1, history: 2, reapable: 1, byLiveness: {}, heavyLane: { executing: 1, nextReady: 1, dormant: 2 }, progressSlo: { oldestWaitMs: 7_200_000, maxNoTransitionMs: 7_200_000 } }}
       workrooms={[
-        { capsuleId: "WC-LIVE", title: "Live rescue intake", status: "working", source: "manual", executorKind: "human", portfolioRole: "manufactureAndDeliver", headBranch: null, pullRequestUrl: null, updatedAt: "2026-08-24T18:00:00.000Z", liveness: "live", isLive: true, isReapable: false, livenessReason: "Lease valid.", trueLivenessAt: "2026-08-24T19:00:00.000Z" },
-        { capsuleId: "WC-OLD", title: "Expired review", status: "working", source: "external-adoption", executorKind: "codex-desktop", portfolioRole: "foundational", headBranch: "fix/old", pullRequestUrl: null, updatedAt: "2026-08-20T18:00:00.000Z", liveness: "lease-expired", isLive: false, isReapable: true, livenessReason: "Lease expired.", trueLivenessAt: "2026-08-20T19:00:00.000Z" },
+        { capsuleId: "WC-LIVE", backlogItemId: "BI-LIVE", title: "Live rescue intake", status: "working", source: "manual", executorKind: "human", portfolioRole: "manufactureAndDeliver", headBranch: null, pullRequestUrl: null, updatedAt: "2026-08-24T18:00:00.000Z", liveness: "live", isLive: true, isReapable: false, livenessReason: "Lease valid.", trueLivenessAt: "2026-08-24T19:00:00.000Z" },
+        { capsuleId: "WC-OLD", backlogItemId: null, title: "Expired review", status: "working", source: "external-adoption", executorKind: "codex-desktop", portfolioRole: "foundational", headBranch: "fix/old", pullRequestUrl: null, updatedAt: "2026-08-20T18:00:00.000Z", liveness: "lease-expired", isLive: false, isReapable: true, livenessReason: "Lease expired.", trueLivenessAt: "2026-08-20T19:00:00.000Z" },
       ]}
     />);
 
@@ -22,7 +22,8 @@ describe("WorkroomInventory", () => {
     expect(html).toContain("Ready");
     expect(html).toContain("Wait");
     expect(html).toContain("2h");
-    expect(html).toContain('href="/workspace/cases/WC-LIVE"');
+    expect(html).toContain('href="/workspace/cases/backlog-item%3ABI-LIVE"');
+    expect(html).toContain('href="/workspace/cases/work-capsule%3AWC-OLD"');
     expect(html).toContain("Manufacture &amp; Deliver");
     expect(html).not.toContain("3 active");
   });
