+++ docs/architecture/build-gate-runbook.md
@@ -16,7 +16,7 @@ TypeScript errors only surface in `next build`, not in `vitest` or IDE checks. R
 
 **Local pre-commit gates.** Pre-commit hook at `.githooks/pre-commit` runs a staged Gitleaks secret scan before bytes enter Git history, then runs `pnpm --filter <affected> typecheck` on `.ts`/`.tsx`/`.mts`/`.cts` commits and rejects on failure. Set once: `git config core.hooksPath .githooks` (auto for new clones via `postinstall`). Manual scans: `pnpm security:secrets:staged` for staged content and `pnpm security:secrets` for the current tree. Emergency bypasses exist for verified false positives only: `DPF_SKIP_SECRET_SCAN=1` and `DPF_SKIP_TYPECHECK=1`; CI still gates the PR.
 
-**Local-CI pre-push gate (default-on).** `pnpm run pregate` leases the shared local-CI sandbox, runs the CI-equivalent runner, and records a SHA-bound evidence record with an expiry; the pre-push hook then REFUSES a push of runtime code without an unexpired passing record for the current branch+SHA. Docs-only diffs (base `DPF_PREPUSH_BASE_REF`, default `origin/main`), delete/tag pushes, and `main` pass without a record. The bypass is recorded, never silent: `DPF_SKIP_PREPUSH_GATE=1 DPF_SKIP_PREPUSH_GATE_REASON="why" git push`, and a PR lacking a gate record for its head SHA needs a `Local-CI-Override:`/`Local-CI-Evidence:` attestation in the PR body to read as merge-ready (`pnpm pr:health` surfaces it). **Pre-PR verification is the sandbox lease — "test on :3000" is only valid AFTER PR/merge/self-upgrade** (§13). Full mechanics, the `Unit Tests` job, and test-runner pins (mobile jest `^29.x`, web/db vitest): [`docs/testing/pre-pr-gate.md`](../testing/pre-pr-gate.md).
+**Local-CI pre-push gate (default-on).** `pnpm run pregate` leases the shared local-CI sandbox, runs the CI-equivalent runner, and records a SHA-bound evidence record with an expiry; the pre-push hook then REFUSES a push of runtime code without an unexpired passing record for the current branch+SHA. Docs-only diffs (base `DPF_PREPUSH_BASE_REF`, default `origin/main`), delete/tag pushes, and `main` pass without a record. A supported bypass is recorded, never silent: after the operator explicitly authorizes the exact run, use `DPF_SKIP_PREPUSH_GATE=1 DPF_SKIP_PREPUSH_GATE_REASON="operator-emergency: <gate, reason, and scope>" git push`. The receipt remains `skipped: true` / `gatePassed: false`; it never converts an unrun or broken gate into a pass. DCO and the cheap deterministic preflight still refuse unsafe changes, and a PR lacking a passing record for its head SHA needs the matching `Local-CI-Override:` attestation in the PR body (`pnpm pr:health` surfaces it). If the failing gate exposes no checked-in audited override, AGENTS.md's refusal rule still requires a stop. **Pre-PR verification is the sandbox lease — "test on :3000" is only valid AFTER PR/merge/self-upgrade** (§13). Full mechanics, the `Unit Tests` job, and test-runner pins (mobile jest `^29.x`, web/db vitest): [`docs/testing/pre-pr-gate.md`](../testing/pre-pr-gate.md).
 
 **Two mandatory pre-push checks the sandbox record does not cover (stuck-PR sweep, PRs #4172/#4173/#4189).** The sandbox gate above verifies a *runtime* record, but two author-fixable failure classes were shipping red because they were only caught in CI, after the push, once the agent had moved on. Both are now enforced host-native in `.githooks/pre-push-gate`, need no sandbox, and cost seconds:
